Update to v24.0.0

+++ Core/include/Acts/Surfaces/detail/VerticesHelper.hpp
@@ -53,7 +53,8 @@ void createSegment(std::vector<vertex_t>& vertices,
                    const vertex_t& offset = vertex_t::Zero(),
                    const transform_t& transform = transform_t::Identity()) {
   // Calculate the number of segments - 1 is the minimum
-  unsigned int segs = std::abs(phi2 - phi1) / (2 * M_PI) * lseg;
+  unsigned int segs =
+      static_cast<unsigned int>(std::abs(phi2 - phi1) / (2 * M_PI) * lseg);
   segs = segs > 0 ? segs : 1;
   ActsScalar phistep = (phi2 - phi1) / segs;
   // Create the segments
